# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AndersSpotify API",
        "description": "Documentación de la API para el proyecto de reproducción musical tipo Spotify",
        "contact": {
            "email": "fernandaurcelay5@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/music/filter": {
            "get": {
                "tags": [
                    "Música"
                ],
                "summary": "Filtrar canciones por categoría, artista o álbum",
                "operationId": "214bf106049ee9b4f695cb815b984349",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "ID de la categoría",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    },
                    {
                        "name": "artist",
                        "in": "query",
                        "description": "ID del artista",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 4
                        }
                    },
                    {
                        "name": "album",
                        "in": "query",
                        "description": "ID del álbum",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Canciones filtradas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Nombre de canción"
                                            },
                                            "file_url": {
                                                "type": "string",
                                                "example": "https://..."
                                            },
                                            "duration": {
                                                "type": "string",
                                                "example": "00:03:45"
                                            },
                                            "play_count": {
                                                "type": "integer",
                                                "example": 500
                                            },
                                            "cover_image": {
                                                "type": "string",
                                                "example": "https://..."
                                            },
                                            "artist": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 4
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Nombre del artista"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "album": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Nombre del álbum"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "category": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Rock"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/music/search": {
            "get": {
                "tags": [
                    "Música"
                ],
                "summary": "Buscar canciones por título",
                "operationId": "b4c0f369268f9a80744f6518676e22f8",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Palabra clave del título de la canción",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "love"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de canciones que coinciden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Love Story"
                                            },
                                            "file_url": {
                                                "type": "string",
                                                "example": "https://..."
                                            },
                                            "cover_image": {
                                                "type": "string",
                                                "example": "https://..."
                                            },
                                            "duration": {
                                                "type": "string",
                                                "example": "00:04:12"
                                            },
                                            "play_count": {
                                                "type": "integer",
                                                "example": 320
                                            },
                                            "likes": {
                                                "type": "integer",
                                                "example": 50
                                            },
                                            "is_featured": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "artist": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Taylor Swift"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "album": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Fearless"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "category": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Pop"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Consulta vacía"
                    }
                }
            }
        },
        "/api/music/{id}": {
            "get": {
                "tags": [
                    "Música"
                ],
                "summary": "Obtener detalle de una canción por ID",
                "operationId": "781486f7c8920f9918770b4a1210f573",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la canción",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de la canción",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Shape of You"
                                        },
                                        "file_url": {
                                            "type": "string",
                                            "example": "https://..."
                                        },
                                        "cover_image": {
                                            "type": "string",
                                            "example": "https://..."
                                        },
                                        "duration": {
                                            "type": "string",
                                            "example": "00:03:54"
                                        },
                                        "play_count": {
                                            "type": "integer",
                                            "example": 1800
                                        },
                                        "likes": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "is_featured": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "artist": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 4
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Ed Sheeran"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "album": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Divide"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "category": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Pop"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Canción no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Usa el token generado tras el login. Formato: Bearer (token)",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}